Move zcl into zspec folder.
