+++ src/adapter/deconz/driver/frameParser.ts
@@ -6,7 +6,7 @@ import PARAM from './constants';
 import { busyQueue, apsBusyQueue, readyToSend, enableRTS, disableRTS, enableRtsTimeout } from './driver';
 import { Request, ReceivedDataResponse, DataStateResponse, Command, ParamMac, ParamPanId, ParamNwkAddr, ParamExtPanId, ParamChannel, ParamChannelMask, ParamPermitJoin, ParamNetworkKey, gpDataInd } from './constants';
 import * as Events from '../../events';
-import {ZclFrame} from '../../../zcl';
+import {ZclFrame} from '../../../zspec/zcl';
 import {logger} from '../../../utils/logger';
 
 const NS = 'zh:deconz:frameparser';
