+++ src/adapter/z-stack/adapter/endpoints.ts
@@ -1,5 +1,5 @@
 import * as Constants from '../constants';
-import {Clusters} from '../../../zcl/index';
+import {Clusters} from '../../../zspec/zcl/index';
 
 const EndpointDefaults: {
     appdeviceid: number;
