+++ src/controller/events.ts
@@ -1,4 +1,4 @@
-import {FrameControl} from "../zcl";
+import {FrameControl} from "../zspec/zcl";
 import {Device, Endpoint} from "./model";
 import {KeyValue} from "./tstype";
