+++ src/zspec/zcl/definition/endpointDeviceType.ts
