+++ src/zspec/zcl/definition/manufacturerCode.ts
