+++ test/controller.test.ts
@@ -5,9 +5,9 @@ import {DeconzAdapter} from '../src/adapter/deconz/adapter';
 import {ZiGateAdapter} from "../src/adapter/zigate/adapter";
 import equals from 'fast-deep-equal/es6';
 import fs from 'fs';
-import { ZclFrame, ZclHeader } from "../src/zcl";
+import { ZclFrame, ZclHeader } from "../src/zspec/zcl";
 import { Device, Group} from "../src/controller/model";
-import * as Zcl from '../src/zcl';
+import * as Zcl from '../src/zspec/zcl';
 import zclTransactionSequenceNumber from '../src/controller/helpers/zclTransactionSequenceNumber';
 import Request from '../src/controller/helpers/request';
 import {Adapter} from '../src/adapter';
@@ -17,7 +17,6 @@ import * as Models from "../src/models";
 import * as Utils from "../src/utils";
 import Bonjour, {BrowserConfig, Service} from 'bonjour-service';
 import {setLogger} from "../src/utils/logger";
-import {createCustomCluster} from "../src/zcl/utils";
 import {BroadcastAddress} from "../src/zspec/enums";
 import ZclTransactionSequenceNumber from "../src/controller/helpers/zclTransactionSequenceNumber";
 const globalSetImmediate = setImmediate;
